Move function to firebase folder
