+++ backend/src/index.js
@@ -19,7 +19,7 @@ const views_2 = __importDefault(require("./customers/views"));
 const swagger_ui_express_1 = __importDefault(require("swagger-ui-express"));
 const api_spec_json_1 = __importDefault(require("../api-spec.json"));
 const database_1 = require("./database");
-const functions_1 = __importDefault(require("./functions"));
+const firebaseFunctions_1 = __importDefault(require("../firebase/firebaseFunctions"));
 const app = (0, express_1.default)();
 // Middleware to parse json request bodies
 app.use(body_parser_1.default.json());
@@ -39,15 +39,15 @@ app.post("/", (req, res) => {
     res.send(req.body);
 });
 /**
- * Create your route here!
- */
-/**
  * Route for creating new user with phone number
  */
 app.get("/create/:number", (req, res) => {
-    (0, functions_1.default)(req.params.number);
+    (0, firebaseFunctions_1.default)(req.params.number);
     res.send("Created new user with number " + req.params.number);
 });
+/**
+ * Create your route here!
+ */
 app.listen(process.env.PORT || 3000, () => __awaiter(void 0, void 0, void 0, function* () {
     console.log("✅ Server is up and running");
     yield (0, database_1.dbConnect)();
